added Exception_Handling

+++ src/main/java/com/EmployeeDirectory/EmployeeManagement/service/serviceImpl/EmplooyeeServiceImpl.java
@@ -1,5 +1,6 @@
 package com.EmployeeDirectory.EmployeeManagement.service.serviceImpl;
 
+import com.EmployeeDirectory.EmployeeManagement.exception.EmployeeNotFoundException;
 import com.EmployeeDirectory.EmployeeManagement.model.EmployeeSchema;
 import com.EmployeeDirectory.EmployeeManagement.repository.EmployeeRepo;
 import com.EmployeeDirectory.EmployeeManagement.service.EmployeeService;
@@ -26,14 +27,33 @@ public class EmplooyeeServiceImpl implements EmployeeService {
 
     public void deleteEmployeeById(int id)
     {
+        if (!employeeRepo.existsById(id)) {
+            throw new EmployeeNotFoundException("Employee with ID " + id + " not found");
+        }
         employeeRepo.deleteById(id);
     }
 
     public void updateEmployeeById(EmployeeSchema newempSchema,int id)
     {
-        if(id==newempSchema.getEmpId())
-        {
-            employeeRepo.save(newempSchema);
+        if (!employeeRepo.existsById(id)) {
+            throw new EmployeeNotFoundException("Employee with ID " + id + " not found");
         }
+//        if(id==newempSchema.getEmpId())
+//        {
+//            employeeRepo.save(newempSchema);
+//        }
+        newempSchema.setEmpId(id);
+        employeeRepo.save(newempSchema);
     }
+    @Override
+    public List<EmployeeSchema> getEmployeesByDepartment(String department) {
+        return employeeRepo.findByDepartment(department);
+    }
+
+    @Override
+    public List<EmployeeSchema> getEmployeesByPosition(String position) {
+        return employeeRepo.findByPosition(position);
+    }
+
+
 }
